requirements.txt
# Core Orchestration
prefect>=3.0.0
prefect-aws>=0.4.0

# AWS SDK
boto3>=1.34.0
botocore>=1.34.0

# Data Processing
pandas>=2.1.0
pyarrow>=14.0.0
numpy>=1.24.0

# API Clients
requests>=2.31.0

# Environment Management
python-dotenv>=1.0.0

# Optional: DataHub Integration
# prefect-datahub>=0.2.0
# datahub>=0.10.0

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

# Optional: Jupyter for exploration
# jupyter>=1.0.0
# ipython>=8.12.0
